Add some more stuff

+++ tools/dump_sys573_gamefs.py
@@ -13,7 +13,6 @@ def get_filename_hash(filename):
 
     return hash & 0xffffffff
 
-
 def decrypt_data(data, input_key):
     def calculate_key(input):
         key = 0
@@ -99,8 +98,14 @@ def decode_lz(input_data):
 
 
 pccard_filenames = [
-    # Files from DDR
+    # General
     "boot/checksum.dat",
+    "boot/psx.bin",
+    "boot/main.exe",
+    "psx.bin",
+    "main.exe",
+
+    # Files from DDR
     "data/fpga/fpga_mp3.bin",
     "data/mp3/mp3_tab.bin",
     "data/tim/wfont/wfont_w.bin",
@@ -140,53 +145,29 @@ pccard_filenames = [
     "ascii_size24.bin",
     "music_info.bin",
     "course_info.bin",
-    "group_list.bin", # Contains a list of files, can be parsed to get fcns listed below
+    "group_list.bin",
+    "jp_title.bin",
     "got11j1b.bin",
     "got11j0b.bin",
-    "got11h1f.bin",
+    "got11hlf.bin",
     "d_res_ns.dat",
     "d_ending.dat",
     "d_title.dat",
     "d_cautio.dat",
     "g_id.dat",
-
-    "system.fcn",
-    "msucheck.fcn",
-    "warning.fcn",
-    "entry.fcn",
-    "playstyl.fcn",
-    "shoprank.fcn",
-    "testmode.fcn",
-    "gamemode.fcn",
-    "ending.fcn",
-    "konamiid.fcn",
-    "card_ent.fcn",
-    "card_res.fcn",
-    "internet.fcn",
-    "scorernk.fcn",
-    "senddata.fcn",
-    "result.fcn",
-    "mus_sel.fcn",
-    "mode_sel.fcn",
-    "caution.fcn",
-    "hitchart.fcn",
-    "share_bg.fcn",
-    "puzzle.fcn",
-    "m_image.fcn",
-    "share.fcn",
-    "bgm0128.fcn",
+    "d_title_aa.dat",
+    "d_cautio_aa.dat",
+    "a.pak",
+    "system.vas",
 
     # Mambo a Gogo
-    "snapshot.img",
-
     "dl/n1/pathtab.bin",
     "data/kanji/eames.cmp",
     "data/spu/system.vas",
-    #"data/mp3/sample/",
-
     "soft/s573/overlay/bin/dbugtest.olb",
     "soft/s573/overlay/bin/gtest.olb",
     "soft/s573/overlay/bin/play.olb",
+    #"data/mp3/sample/",
     # "soft/s573/overlay/bin/fpga_mp3.olb", # There's a fpga_mp3_new.bin somewhere in here but I have no idea what filename it's using
 ]
 
@@ -363,6 +344,17 @@ def parse_mdb_filenames(data):
 
 
 # start GFDM data
+def parse_group_list_filenames(data):
+    hash_list_add = {}
+
+    for i in range(len(data) // 0x20):
+        filename = data[i*0x20:(i+1)*0x20].decode('ascii').strip('\0')
+
+        path = "%s.fcn" % (filename)
+        hash_list_add[get_filename_hash(path)] = path
+
+    return hash_list_add
+
 # end GFDM data
 
 # start Mambo a Gogo data
@@ -502,7 +494,7 @@ def get_file_data(fileinfo, enckey=None):
     if data and fileinfo['flag_comp'] == 1:
         data = decode_lz(data)
 
-    return data
+    return bytearray(data)
 
 
 if __name__ == "__main__":
@@ -549,10 +541,26 @@ if __name__ == "__main__":
                 if hash_list[fileinfo['filename_hash']] == "data/mdb/mdb.bin":
                     hash_list.update(parse_mdb_mambo_filenames(get_file_data(fileinfo, args.key)))
 
+            elif args.type == "gfdm":
+                if hash_list[fileinfo['filename_hash']] == "group_list.bin":
+                    hash_list.update(parse_group_list_filenames(get_file_data(fileinfo, args.key)))
+
+    for idx, fileinfo in enumerate(files):
+        if fileinfo['filename_hash'] in hash_list:
+            if args.type in ["ddr", "mambo"]:
+                if hash_list[fileinfo['filename_hash']] == "data/tex/rembind.bin":
+                    hash_list.update(parse_rembind_filenames(get_file_data(fileinfo, args.key)))
+
+            if args.type == "ddr":
+                if hash_list[fileinfo['filename_hash']] == "data/mdb/mdb.bin":
+                    hash_list.update(parse_mdb_filenames(get_file_data(fileinfo, args.key)))
+
+            elif args.type == "mambo":
+                if hash_list[fileinfo['filename_hash']] == "data/mdb/mdb.bin":
+                    hash_list.update(parse_mdb_mambo_filenames(get_file_data(fileinfo, args.key)))
+
     for idx, fileinfo in enumerate(files):
         output_filename = "output_%04d.bin" % idx
-
-        print(fileinfo)
 
         if fileinfo['offset'] == 0x7c0000:
             print("%08x" % fileinfo['offset'], fileinfo)
@@ -563,23 +571,20 @@ if __name__ == "__main__":
             if output_filename.startswith("/"):
                 output_filename = "_" + output_filename[1:]
 
-            output_filename = os.path.join(args.output, output_filename)
-
-            print("Extracting", output_filename)
-
-            filepath = os.path.dirname(output_filename)
-            if not os.path.exists(filepath):
-                os.makedirs(filepath)
-
-            with open(output_filename, "wb") as outfile:
-                data = get_file_data(fileinfo, args.key)
-                outfile.write(data)
-
         else:
-            print("Found unknown hash: %08x" % fileinfo['filename_hash'])
+            print("Unknown hash %08x" % fileinfo['filename_hash'], output_filename)
 
-            output_filename = os.path.join(args.output, output_filename)
+        output_filename = os.path.join(args.output, output_filename)
 
-            with open(output_filename, "wb") as outfile:
-                data = get_file_data(fileinfo, args.key)
-                outfile.write(data)
+        if os.path.exists(output_filename):
+            continue
+
+        filepath = os.path.dirname(output_filename)
+        if not os.path.exists(filepath):
+            os.makedirs(filepath)
+
+        print(fileinfo)
+        print("Extracting", output_filename)
+        with open(output_filename, "wb") as outfile:
+            data = get_file_data(fileinfo, args.key)
+            outfile.write(data)
